Enable static/shared libs in same build process.
This commit allows for building of static and/or shared object library
files in the same build process. These are controlled by cmake flags

-DBUILD_STATIC_LIBS=ON|OFF

-DBUILD_SHARED_LIBS=ON|OFF

The default for both is ON.

Further modifications were necessary to the other CMake config scripts
to ensure successful runs in the new build workflow.

This commit references issue #408