Cleanup of aecpack.c
General clean up of aecpack.c.

In the function, logic has been added to default values for the
AEC/CCSDS parameters if they are set to zero in the input DRS
template For nbits, added logic to round up the value to
the next base 2 integer.  For example, nbits set to 12 in the scaling
part of the code will be rounded to 16.